requirements.txt
wheel~=0.37.1
fastapi~=0.90.0
pip~=23.0
pydantic~=1.10.4
SQLAlchemy~=2.0.2
uvicorn~=0.20.0
pyyaml
dataclass-factory